Moved interfaces to interface folder
