+++ AssetTracking/AssetTracking/Interfaces/IOfficeBookRepository.cs
@@ -3,7 +3,7 @@ using Microsoft.Graph;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 
-namespace AssetTracking.Repositories
+namespace AssetTracking.Interfaces
 {
     public interface IOfficeBookRepository
     {
